Extract sample browser rows
